+++ CHANGELOG.md
@@ -9,12 +9,16 @@
   `HTML_TOC` object to the `Markdown` object's constructor since
   some languages rely on the sharp to comment code.
 
+  Fixes #451.
+
 * Allow passing `Range` objects to the `nesting_level` option to have
   a higher level of customization for table of contents:
 
   ~~~ruby
   Redcarpet::Render::HTML_TOC.new(nesting_level: 2..5)
   ~~~
+
+  Fixes #519.
 
 ## Version 3.4.0